pyproject.toml
[tool.poetry]
name = "ocr-reader"
version = "0.1.0"
description = "The command line tool extract text from image."
authors = ["Nhu Son"]

[tool.poetry.dependencies]
python = "3.7.6"
click = "^7.1.2"
pytesseract = "^0.3.4"
Pillow = "^7.2.0"
pdf2image = "^1.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
